@@ -1,48 +1,50 @@
 {
-  "name": "ionic-app-base",
+  "name": "tunnel-app",
+  "version": "0.0.1",
   "author": "Ionic Framework",
   "homepage": "http://ionicframework.com/",
   "private": true,
-  "version": "1.0.0",
   "scripts": {
     "clean": "ionic-app-scripts clean",
     "build": "ionic-app-scripts build",
+    "lint": "ionic-app-scripts lint",
     "ionic:build": "ionic-app-scripts build",
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "2.2.1",
-    "@angular/compiler": "2.2.1",
-    "@angular/compiler-cli": "2.2.1",
-    "@angular/core": "2.2.1",
-    "@angular/forms": "2.2.1",
-    "@angular/http": "2.2.1",
-    "@angular/platform-browser": "2.2.1",
-    "@angular/platform-browser-dynamic": "2.2.1",
-    "@angular/platform-server": "2.2.1",
-    "@ionic/storage": "1.1.7",
-    "angular2-baidu-map": "2.4.1",
-    "baidu-map": "^0.1.4",
-    "ionic-angular": "2.0.0-rc.5",
-    "ionic-native": "2.2.11",
+    "@angular/common": "4.1.3",
+    "@angular/compiler": "4.1.3",
+    "@angular/compiler-cli": "4.1.3",
+    "@angular/core": "4.1.3",
+    "@angular/forms": "4.1.3",
+    "@angular/http": "4.1.3",
+    "@angular/platform-browser": "4.1.3",
+    "@angular/platform-browser-dynamic": "4.1.3",
+    "@ionic-native/core": "3.10.2",
+    "@ionic-native/splash-screen": "3.10.2",
+    "@ionic-native/status-bar": "3.10.2",
+    "@ionic/storage": "2.0.1",
+    "angular2-baidu-map": "^2.4.1",
+    "ionic-angular": "3.4.2",
+    "ionic-native": "^2.9.0",
     "ionicons": "3.0.0",
-    "jquery": "^3.1.1",
-    "leaflet": "^0.7.7",
-    "lodash": "^4.15.0",
-    "pouchdb": "^5.4.5",
-    "pouchdb-find": "^0.10.3",
-    "reflect-metadata": "^0.1.3",
-    "rxjs": "5.0.0-beta.12",
-    "sw-toolbox": "3.4.0",
-    "zone.js": "0.6.26"
+    "jquery": "^3.2.1",
+    "leaflet": "^1.0.3",
+    "lodash": "^4.17.4",
+    "pouchdb": "^6.2.0",
+    "pouchdb-find": "^6.2.0",
+    "rxjs": "5.4.0",
+    "sw-toolbox": "3.6.0",
+    "zone.js": "0.8.12"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "1.0.0",
+    "@ionic/app-scripts": "1.3.7",
     "@ionic/cli-plugin-cordova": "1.4.0",
     "@ionic/cli-plugin-ionic-angular": "1.3.1",
-    "@types/jquery": "^2.0.39",
-    "@types/lodash": "^4.14.50",
-    "@types/pouchdb": "^5.4.28",
-    "typescript": "2.0.9"
-  }
+    "@types/jquery": "^3.2.3",
+    "@types/lodash": "^4.14.66",
+    "@types/pouchdb": "^6.1.1",
+    "typescript": "2.3.3"
+  },
+  "description": "An Ionic project"
 }
