@@ -1,3 +1,4 @@
+import { MenuController } from 'ionic-angular/umd';
 import { ToggleMenu } from '../../../shared/components/toggle-menu/toggle-menu.component';
 import { Component, AfterViewInit } from '@angular/core';
 
@@ -10,7 +11,7 @@ interface Autodest{
   templateUrl: './manyou.component.html',
   styles: ['./manyou.component.scss']
 })
-export class ManyouComponent extends ToggleMenu implements AfterViewInit {
+export class ManyouComponent implements AfterViewInit {
 
   ngAfterViewInit() {
     
